+++ src/app/components/DynamicForm/styles.ts
@@ -1,5 +1,5 @@
 export const dynamicFormStyles = {
-  container: 'w-full max-w-2xl mx-auto p-4 space-y-6',
+  container: 'w-full  p-4 space-y-6',
   heading: 'text-2xl font-bold',
   form: 'space-y-6',
   actions: 'flex flex-col sm:flex-row justify-center sm:justify-end gap-4 w-full',