+++ src/app/main-page/accounting/personalInvoices/coments/page.tsx
@@ -1,0 +1,4 @@
+const PersonalInvoicesComents = ()=>{
+    return (<>Comments</>)
+}
+export default PersonalInvoicesComents;